refactor(nika-check,nika-builtin): la masse de test etait facturee au budget de PROD
DEUX PORTES DE TAILLE ONT ROUGI SUR CETTE BRANCHE, et aucune des deux ne
parlait vraiment d elle. Les 26 lignes de production qu elle ajoute ont
seulement REVELE que les deux compteurs facturaient du test.

LA LOI DES DEUX COMPTEURS
  crate-size (15k)  soustrait #[cfg(test)] ET les fichiers dont le
                    BASENAME est tests.rs · « a tests.rs sibling is 100%
                    test code », dit son propre commentaire
  file-loc  (1500)  ne soustrait RIEN · c est un wc -l

① nika-check · 15002 → 14996
permits_fit_tests.rs est 1097 lignes de batterie, et son en-tete le dit ·
« The permit-fit families under test ». Il etait declare par
`#[path = "permits_fit_tests.rs"]`, donc son basename n est pas tests.rs,
donc le compteur le lisait comme de la PRODUCTION · 13 lignes (SPDX,
doc de module, use super::*) facturees au budget des 15k. Il devient
permits_fit/tests.rs · exactement la forme que son voisin hints/tests.rs
portait deja dans cette meme crate.

② nika-builtin · data.rs 1648 → 650
646 lignes de production portaient 1000 lignes de batterie, et c est
file-loc qui rougit · lui ne soustrait pas cfg(test), donc un module sain
traversait le plafond de 1500 par sa seule masse de test. Les deux modules
(tests · expression_boundary) partent dans data/tests.rs. Meme nom de
fichier, donc les DEUX compteurs le soustraient desormais.

CE QUI N A PAS ETE FAIT, ET POURQUOI
Pas de marqueur LOC-EXEMPT · les trois raisons admises sont codegen,
lookup-table, enum-mega, et aucune ne decrit ce fichier. Une exemption
inventee pour un fichier qu on vient de faire grossir, c est un vert
achete. Pas de rognage de commentaires non plus · le compteur compte les
lignes de commentaire, donc « descendre » en coupant du raisonnement
serait une mesure amelioree et un code appauvri.

⚠️ LE MUR RESTE LA, ET IL EST DATE
nika-check tient a 14996 sur 15000 · QUATRE lignes. Le mur ne vient pas
de cette branche · main y etait deja a 14976. La prochaine addition
serieuse dans cette crate demandera la descente (ADR-022 / ADR-024), pas
un autre deplacement de fichier. C est dit ici pour que personne ne le
decouvre en urgence.

Verifie · 360 tests nika-builtin avant, 360 apres, meme compte · clippy
--all-targets -D warnings propre sur les deux crates · file-loc et
crate-size rendent YELLOW, plus aucun RED.
