+++ crates/nika-check/src/permits_fit/tests.rs
